Remove some questionable trait impls
